+++ src/interfaces/maps_api_options.ts
@@ -1,0 +1,4 @@
+export interface MapsListOptions {
+  limit?: number;
+  accountHandle?: string;
+}